Extract RopeText
